Refactor login form not to use custom form since it doesn't need to
