+++ src/auth/components/LoginPage/types.ts
@@ -1,0 +1,4 @@
+export interface LoginFormData {
+  email: string;
+  password: string;
+}
